📝 Modify(#79): 인터셉터 instance로 옮기기
